requirements.txt
django-dbbackup
django-role-permissions
apt-get install libapache2-mod-wsgi
apt-get install python-dev
apt-get install libapache2-mod-python
